+++ src/Elastic.Channels/BufferedChannelBase.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
+using Elastic.Channels.Buffers;
 
 namespace Elastic.Channels;
