+++ src/Elastic.Channels/ResponseItemsBufferedChannelBase.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Elastic.Channels.Buffers;
 
 namespace Elastic.Channels;
